Use double universally for precision/ease

+++ src/main/java/dev/salmonllama/snack/Constants.java
@@ -1,11 +1,11 @@
 package dev.salmonllama.snack;
 
 public class Constants {
-    public static final float DRINK_PRICE = 1.50F;
+    public static final double DRINK_PRICE = 1.50F;
 
-    public static final float CHIP_PRICE = 1.00F;
+    public static final double CHIP_PRICE = 1.00F;
 
     // More items possible by declaring them here.
 
-    public static final float SALES_TAX = 0.05F; // 5%
+    public static final double SALES_TAX = 0.05F; // 5%
 }
